Update python examples
