+++ python/tests/bullet_Talos.py
@@ -9,7 +9,7 @@ import pybullet_data
 import pybullet as p  # PyBullet simulator
 import numpy as np
 from scipy.spatial.transform import Rotation as R
-import configuration as conf
+import configurationWalk as conf
 # import os
